simulator/relay-mqtt-webserver/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
paho-mqtt==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0